third_party/blink/web_tests/platform/generic/fast/forms/textarea/textarea-initial-caret-position-expected.txt

This tests the initial caret position for textareas.

Bug 11746: REGRESSION(r14931): Outlook Web Access incorrectly positions the insertion point when replying to e-mail
http://bugs.webkit.org/show_bug.cgi?id=11746

PASS ta.selectionStart is 0
PASS ta.selectionEnd is 0
PASS successfullyParsed is true

TEST COMPLETE
